Use LDFLAGS when compiling zerovm
This lets the compiler find libvalidator correctly when it is
installed under PREFIX/lib and zerovm is compilsed using the same
PREFIX setting.

This removes the "need" to run 'sudo make install' when installing the
validator, e.g., the library can be installed in a custom location be
found without further hacking of the zerovm Makefile.